feat: deletion animation

+++ lib/infra/views/elements/flash_card.dart
@@ -2,6 +2,7 @@ import 'package:flash_clouds_app/domain/entities/card_entity.dart';
 import 'package:flash_clouds_app/infra/data_structures/cards_list.dart';
 import 'package:flash_clouds_app/infra/local_db/cards_repository.dart';
 import 'package:flip_card/flip_card.dart';
+import 'package:flip_card/flip_card_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -17,6 +18,18 @@ class FlashCard extends StatefulWidget {
 }
 
 class _FlashCardState extends State<FlashCard> {
+  late FlipCardController _controller;
+  Color _cardColor = Colors.white;
+  double _opacity = 1.0;
+  Duration _opacityAnimationDuration = const Duration(seconds: 1);
+  bool _enableControls = true;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = FlipCardController();
+  }
+
   @override
   Widget build(BuildContext context) {
     String front = widget.cardEntity.front;
@@ -27,12 +40,18 @@ class _FlashCardState extends State<FlashCard> {
       back = widget.cardEntity.front;
     }
 
-    return FlipCard(
-      // Fill the back side of the card to make in the same size as the front.
-      fill: front.length > back.length ? Fill.fillBack : Fill.fillFront,
-      direction: FlipDirection.HORIZONTAL, // default
-      front: _cardSide(context, front),
-      back: _cardSide(context, back),
+    return AnimatedOpacity(
+      opacity: _opacity,
+      duration: _opacityAnimationDuration,
+      child: FlipCard(
+        // Fill the back side of the card to make in the same size as the front.
+        fill: front.length > back.length ? Fill.fillBack : Fill.fillFront,
+        direction: FlipDirection.HORIZONTAL, // default
+        front: _cardSide(context, front),
+        back: _cardSide(context, back),
+        controller: _controller,
+        flipOnTouch: _enableControls,
+      ),
     );
   }
 
@@ -57,11 +76,24 @@ class _FlashCardState extends State<FlashCard> {
       children: [
         IconButton(
             onPressed: () async {
+              if (!_enableControls) {
+                return;
+              }
+              setState(() {
+                _enableControls = false;
+              });
+
+              // without this, after deleting a flipped card, next card in list gets flipped instantly
+              await _animateDeletion();
+
               widget.cardEntity.delete();
               List<CardEntity?> cards =
                   await CardsRepository().getAllSortByDate();
-
               Provider.of<CardsList>(context, listen: false).updateList(cards);
+
+              setState(() {
+                _enableControls = true;
+              });
             },
             icon: const Icon(
               Icons.delete_outlined,
@@ -94,7 +126,7 @@ class _FlashCardState extends State<FlashCard> {
 
   BoxDecoration _cardDecoration() {
     return BoxDecoration(
-        color: Colors.white,
+        color: _cardColor,
         border: Border.all(
           width: 1.0,
           color: Colors.blueGrey.withOpacity(0.3),
@@ -108,5 +140,23 @@ class _FlashCardState extends State<FlashCard> {
             offset: const Offset(0, 3),
           )
         ]);
+  }
+
+  Future<void> _animateDeletion() async {
+    // without this, after deleting a flipped card, next card in list gets flipped instantly
+    setState(() {
+      _cardColor = Colors.redAccent;
+      _opacity = 0.0;
+      _opacityAnimationDuration = const Duration(seconds: 1);
+    });
+    if (!_controller.state!.isFront) {
+      _controller.toggleCard();
+    }
+    await Future.delayed(const Duration(seconds: 1));
+    setState(() {
+      _cardColor = Colors.white;
+      _opacity = 1.0;
+      _opacityAnimationDuration = Duration.zero;
+    });
   }
 }
